requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
confluent-kafka==2.1.1
dataclasses-json==0.5.13
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.1
filelock==3.12.2
flatbuffers==23.5.26
frozendict==2.3.8
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
html5lib==1.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.16.0
jupyter==1.0.0
jupyter-client==7.3.5
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.13.1
langchain==0.0.239
langsmith==0.0.14
libclang==16.0.6
lxml==4.9.1
Markdown==3.4.4
MarkupSafe==2.1.1
marshmallow==3.20.1
matplotlib-inline==0.1.6
mistune==2.0.4
mpmath==1.3.0
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.5.0
nest-asyncio==1.5.5
networkx==3.1
nltk==3.8.1
notebook==6.4.12
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openapi-schema-pydantic==1.2.4
opt-einsum==3.3.0
packaging==21.3
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkgutil-resolve-name==1.3.10
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==4.24.2
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.10.1
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.19
stack-data==0.5.0
sympy==1.12
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado==6.2
tqdm==4.66.1
traitlets==5.4.0
transformers==4.32.0
typing-extensions==4.5.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==2.3.7
widgetsnbextension==4.0.3
wrapt==1.15.0
yarl==1.9.2
yfinance==0.2.22
zipp==3.8.1
